Thorns 1.6
This has been a headache and a frustration. It is not perfect and still has a lot more to work on. I added a combat sequence and simplified the plant growth scene. Also, I realized that coding is a lot like being addicted to drugs. You think about it whenever you're not doing it. Sometimes you hate yourself while you're doing it. But then when you get it right you feel very good, maybe you get an adrenaline rush, then you think about how far you can go with it next. And the cycle continues.